add gemini service and seperate the xai to seperate route file and add get attackdetection xai explaination route
