+++ pyproject.toml
@@ -5,7 +5,7 @@ description = "Default template for PDM package"
 authors = [
     {name = "Hamiz5401", email = "kittiporn132@gmail.com"},
 ]
-dependencies = ["fastapi>=0.115.6", "pre-commit>=4.0.1", "uvicorn>=0.34.0", "sqlalchemy>=2.0.36", "xgboost>=2.1.3", "pandas>=2.2.3", "pyarrow>=18.1.0", "python-multipart>=0.0.19", "pydantic-settings>=2.7.1", "boto3>=1.35.94", "redis>=5.2.1", "scikit-learn>=1.6.0", "shap>=0.46.0", "matplotlib>=3.9.4"]
+dependencies = ["fastapi>=0.115.6", "pre-commit>=4.0.1", "uvicorn>=0.34.0", "sqlalchemy>=2.0.36", "xgboost>=2.1.3", "pandas>=2.2.3", "pyarrow>=18.1.0", "python-multipart>=0.0.19", "pydantic-settings>=2.7.1", "boto3>=1.35.94", "redis>=5.2.1", "scikit-learn>=1.6.0", "shap>=0.46.0", "matplotlib>=3.9.4", "google-genai>=1.0.0"]
 requires-python = "==3.9.*"
 readme = "README.md"
 license = {text = "MIT"}
